feat(gateway): split manager concerns
