+++ crates/lab/src/dispatch/gateway/manager.rs
@@ -1,25 +1,19 @@
-use std::collections::{BTreeMap, BTreeSet, HashMap};
+use std::collections::{BTreeMap, BTreeSet};
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::sync::{Arc, Mutex as StdMutex, OnceLock};
 use std::time::Duration;
 
 use arc_swap::ArcSwapOption;
-#[cfg(unix)]
-use nix::errno::Errno;
 use tokio::sync::{Mutex, RwLock};
 use tokio::task::AbortHandle;
 use tokio::time::Instant;
-use url::Url;
 
 use crate::config::{
-    LabConfig, ToolSearchConfig, UpstreamConfig, UpstreamOauthConfig, UpstreamOauthMode,
-    UpstreamOauthRegistration, backup_env, env_is_up_to_date, write_env,
+    LabConfig, ToolSearchConfig, UpstreamConfig, backup_env, env_is_up_to_date, write_env,
 };
 use crate::dispatch::clients::SharedServiceClients;
 use crate::dispatch::error::ToolError;
-use crate::dispatch::gateway::oauth::UpstreamOauthStatusView;
-use crate::dispatch::redact::{redact_stdio_args, redact_stdio_value, redact_url};
 use crate::dispatch::upstream::pool::{
     UpstreamCachedSummary, UpstreamPool, in_process_upstream_name,
 };
@@ -27,11 +21,6 @@ use crate::dispatch::upstream::types::UpstreamRuntimeOwner;
 use crate::oauth::upstream::cache::OauthClientCache;
 use crate::oauth::upstream::encryption::EncryptionKey;
 use crate::oauth::upstream::manager::UpstreamOauthManager;
-use crate::oauth::upstream::types::{BeginAuthorization, OauthError};
-#[cfg(target_os = "linux")]
-use crate::process::unix::read_cmdline;
-#[cfg(unix)]
-use crate::process::unix::{pid_is_alive, terminate_sigkill};
 use crate::registry::ToolRegistry;
 use lab_apis::extract::types::ServiceCreds;
 
@@ -39,19 +28,18 @@ use super::config::{
     default_gateway_bearer_env_name, insert_upstream, load_gateway_config, remove_upstream,
     update_upstream, validate_bearer_token_env_name, validate_tool_search, write_gateway_config,
 };
+use super::config_mutation::{ensure_stdio_admin_ack, read_env_values, values_to_service_creds};
 use super::index::{SearchHit, ToolIndex};
 use super::params::GatewayUpdatePatch;
+use super::projection::*;
+pub use super::runtime::GatewayRuntimeHandle;
+use super::runtime::runtime_origin_tag;
 use super::service_catalog::service_meta;
 use super::types::{
-    CatalogChangeNotifier, GatewayCatalogDiff, GatewayConfigView, GatewayRuntimeView,
-    GatewayToolExposureRowView, GatewayView, ServiceConfigFieldView, ServiceConfigView,
-    VirtualServerMcpPolicyView,
+    CatalogChangeNotifier, GatewayCatalogDiff, GatewayRuntimeView, GatewayToolExposureRowView,
+    GatewayView, ServiceConfigView, VirtualServerMcpPolicyView,
 };
-use super::view_models::{
-    ServerConfigSummaryView, ServerView, SurfaceStateView, SurfaceStatesView,
-};
-use super::virtual_servers::{VirtualServerRecord, VirtualServerSource};
-use crate::tui::events::ServiceHealth;
+use super::view_models::ServerView;
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub struct GatewayCatalogSnapshot {
@@ -127,71 +115,21 @@ pub struct GatewayToolSearchResult {
     pub input_schema: Option<serde_json::Value>,
 }
 
-fn tool_error_from_oauth(error: OauthError) -> ToolError {
-    ToolError::Sdk {
-        sdk_kind: error.kind().to_string(),
-        message: error.to_string(),
-    }
-}
-
-#[derive(Clone, Default)]
-pub struct GatewayRuntimeHandle {
-    pool: Arc<RwLock<Option<Arc<UpstreamPool>>>>,
-}
-
-impl GatewayRuntimeHandle {
-    pub async fn current_pool(&self) -> Option<Arc<UpstreamPool>> {
-        self.pool.read().await.clone()
-    }
-
-    pub async fn swap(&self, pool: Option<Arc<UpstreamPool>>) {
-        *self.pool.write().await = pool;
-    }
-}
-
 #[derive(Clone)]
 pub struct GatewayManager {
-    path: PathBuf,
-    runtime: GatewayRuntimeHandle,
-    config: Arc<RwLock<LabConfig>>,
-    config_mutation: Arc<Mutex<()>>,
+    pub(super) path: PathBuf,
+    pub(super) runtime: GatewayRuntimeHandle,
+    pub(super) config: Arc<RwLock<LabConfig>>,
+    pub(super) config_mutation: Arc<Mutex<()>>,
     service_clients: Option<SharedServiceClients>,
     notifier: Option<CatalogChangeNotifier>,
-    oauth_client_cache: Option<OauthClientCache>,
-    upstream_oauth_managers: Option<Arc<dashmap::DashMap<String, UpstreamOauthManager>>>,
+    pub(super) oauth_client_cache: Option<OauthClientCache>,
+    pub(super) upstream_oauth_managers: Option<Arc<dashmap::DashMap<String, UpstreamOauthManager>>>,
     /// Resources needed to build transient OAuth managers for probed upstreams.
-    oauth_sqlite: Option<lab_auth::sqlite::SqliteStore>,
-    oauth_key: Option<EncryptionKey>,
-    oauth_redirect_uri: Option<Arc<String>>,
+    pub(super) oauth_sqlite: Option<lab_auth::sqlite::SqliteStore>,
+    pub(super) oauth_key: Option<EncryptionKey>,
+    pub(super) oauth_redirect_uri: Option<Arc<String>>,
     tool_indexes: Arc<dashmap::DashMap<String, GatewayToolIndexState>>,
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-struct PersistedGatewayRuntimeState {
-    #[serde(default)]
-    reconciled_at_epoch_secs: Option<u64>,
-    #[serde(default)]
-    entries: Vec<PersistedGatewayRuntimeEntry>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-struct PersistedGatewayRuntimeEntry {
-    upstream: String,
-    pid: u32,
-    #[serde(default)]
-    pgid: Option<u32>,
-    #[serde(default)]
-    started_at_epoch_secs: Option<u64>,
-    #[serde(default)]
-    observed_at_epoch_secs: u64,
-    #[serde(default)]
-    origin: Option<String>,
-    #[serde(default)]
-    owner: Option<crate::dispatch::gateway::types::GatewayRuntimeOwnerView>,
-    #[serde(default)]
-    transport: Option<String>,
-    #[serde(default)]
-    target: Option<String>,
 }
 
 impl GatewayManager {
@@ -271,291 +209,20 @@ impl GatewayManager {
         self.oauth_client_cache.clone()
     }
 
-    pub async fn oauth_upstream_configs(&self) -> Vec<UpstreamConfig> {
-        self.config
-            .read()
-            .await
-            .upstream
-            .iter()
-            .filter(|upstream| upstream.oauth.is_some())
-            .cloned()
-            .collect()
-    }
-
-    pub async fn oauth_upstream_config(&self, upstream_name: &str) -> Option<UpstreamConfig> {
-        self.config
-            .read()
-            .await
-            .upstream
-            .iter()
-            .find(|upstream| upstream.name == upstream_name && upstream.oauth.is_some())
-            .cloned()
-    }
-
     /// Probe `url` for OAuth support via RFC 8414 AS metadata discovery.
     ///
     /// On success, registers a transient `UpstreamOauthManager` (Dynamic strategy)
     /// keyed by the URL hostname so subsequent `begin_upstream_authorization` calls
     /// work without requiring a static config entry.
-    pub async fn probe_upstream_oauth(
-        &self,
-        url: &str,
-    ) -> Result<crate::dispatch::gateway::oauth::ProbeResult, ToolError> {
-        use rmcp::transport::AuthorizationManager;
-        let started = std::time::Instant::now();
-        let redacted_url = redact_url(url);
-
-        // SSRF validation (synchronous DNS) — must run in spawn_blocking.
-        // Also enforces https-only and rejects RFC 1918, loopback, and link-local.
-        let url_for_check = url.to_string();
-        tokio::task::spawn_blocking(move || {
-            crate::dispatch::marketplace::validate_registry_url(&url_for_check)
-        })
-        .await
-        .map_err(|e| ToolError::internal_message(format!("SSRF validation task panicked: {e}")))
-        .inspect_err(|_| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "probe",
-                url = %redacted_url,
-                kind = "ssrf_blocked",
-                "upstream oauth probe: SSRF validation task error"
-            );
-        })??;
-
-        let parsed = Url::parse(url).map_err(|_| ToolError::Sdk {
-            sdk_kind: "invalid_param".to_string(),
-            message: format!("invalid upstream URL: {url}"),
-        })?;
-
-        // Use the raw hostname (with dots) as the DashMap key so that
-        // `api.example.com` and `api-example.com` do not collide.
-        let name = parsed.host_str().unwrap_or("upstream").to_string();
-
-        tracing::info!(
-            service = "upstream_oauth",
-            action = "probe",
-            upstream = %name,
-            url = %redacted_url,
-            "upstream oauth probe: connecting"
-        );
-
-        let auth_manager = AuthorizationManager::new(url).await.map_err(|e| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "probe",
-                upstream = %name,
-                url = %redacted_url,
-                kind = "network_error",
-                error = %e,
-                elapsed_ms = started.elapsed().as_millis(),
-                "upstream oauth probe: connection failed"
-            );
-            ToolError::Sdk {
-                sdk_kind: "network_error".to_string(),
-                message: format!("failed to connect to upstream: {e}"),
-            }
-        })?;
-
-        let metadata = match auth_manager.discover_metadata().await {
-            Ok(m) => {
-                tracing::info!(
-                    service = "upstream_oauth",
-                    action = "probe",
-                    upstream = %name,
-                    url = %redacted_url,
-                    issuer = m.issuer.as_deref().unwrap_or("<none>"),
-                    supports_dynamic_registration = m.registration_endpoint.is_some(),
-                    scopes = ?m.scopes_supported,
-                    elapsed_ms = started.elapsed().as_millis(),
-                    "upstream oauth probe: OAuth metadata discovered"
-                );
-                m
-            }
-            Err(e) => {
-                tracing::info!(
-                    service = "upstream_oauth",
-                    action = "probe",
-                    upstream = %name,
-                    url = %redacted_url,
-                    reason = %e,
-                    elapsed_ms = started.elapsed().as_millis(),
-                    "upstream oauth probe: no OAuth metadata found"
-                );
-                return Ok(crate::dispatch::gateway::oauth::ProbeResult {
-                    upstream: name,
-                    url: url.to_string(),
-                    oauth_discovered: false,
-                    issuer: None,
-                    scopes: None,
-                    registration_strategy: None,
-                });
-            }
-        };
-
-        let supports_dynamic = metadata.registration_endpoint.is_some();
-        let strategy = if supports_dynamic {
-            "dynamic"
-        } else {
-            "client_metadata_document"
-        };
-
-        // Check each prerequisite independently so the error names only what's missing.
-        if self.oauth_key.is_none()
-            || self.oauth_sqlite.is_none()
-            || self.oauth_redirect_uri.is_none()
-        {
-            let missing: Vec<&str> = [
-                self.oauth_key
-                    .is_none()
-                    .then_some("LAB_OAUTH_ENCRYPTION_KEY"),
-                // redirect_uri derives from LAB_PUBLIC_URL; missing key is the same root cause
-                self.oauth_redirect_uri
-                    .is_none()
-                    .then_some("LAB_PUBLIC_URL"),
-            ]
-            .into_iter()
-            .flatten()
-            .collect();
-            let message = format!(
-                "upstream OAuth not configured — set {} to enable it",
-                missing.join(" and ")
-            );
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "probe",
-                upstream = %name,
-                kind = "not_configured",
-                elapsed_ms = started.elapsed().as_millis(),
-                %message,
-                "upstream oauth probe: oauth resources not configured"
-            );
-            return Err(ToolError::Sdk {
-                sdk_kind: "not_configured".to_string(),
-                message,
-            });
-        }
-
-        match (
-            self.upstream_oauth_managers.as_ref(),
-            self.oauth_sqlite.as_ref(),
-            self.oauth_key.as_ref(),
-            self.oauth_redirect_uri.as_ref(),
-        ) {
-            (Some(managers), Some(sqlite), Some(key), Some(redirect_uri)) => {
-                if managers.contains_key(&name) {
-                    tracing::info!(
-                        service = "upstream_oauth",
-                        action = "probe",
-                        upstream = %name,
-                        elapsed_ms = started.elapsed().as_millis(),
-                        "upstream oauth probe: reusing existing manager"
-                    );
-                } else {
-                    let registration = if supports_dynamic {
-                        UpstreamOauthRegistration::Dynamic
-                    } else {
-                        // No RFC 7591 dynamic registration — use the Client ID Metadata
-                        // Document (CIMD) approach: the lab's own metadata-document URL
-                        // acts as the client_id. Derive it from the redirect_uri origin.
-                        let metadata_doc_url = Url::parse(redirect_uri.as_str())
-                            .ok()
-                            .map(|mut u| {
-                                u.set_path("/.well-known/oauth-client");
-                                u.set_query(None);
-                                u.set_fragment(None);
-                                u.to_string()
-                            })
-                            .unwrap_or_default();
-                        UpstreamOauthRegistration::ClientMetadataDocument {
-                            url: metadata_doc_url,
-                        }
-                    };
-                    let config = UpstreamConfig {
-                        enabled: true,
-                        name: name.clone(),
-                        url: Some(url.to_string()),
-                        bearer_token_env: None,
-                        command: None,
-                        args: vec![],
-                        proxy_resources: false,
-                        proxy_prompts: false,
-                        expose_tools: None,
-                        expose_resources: None,
-                        expose_prompts: None,
-                        oauth: Some(UpstreamOauthConfig {
-                            mode: UpstreamOauthMode::AuthorizationCodePkce,
-                            registration,
-                            scopes: metadata.scopes_supported.clone(),
-                        }),
-                        tool_search: ToolSearchConfig::default(),
-                    };
-                    let manager = UpstreamOauthManager::new(
-                        sqlite.clone(),
-                        key.clone(),
-                        config,
-                        redirect_uri.as_ref().clone(),
-                    );
-                    managers.insert(name.clone(), manager);
-                    tracing::info!(
-                        service = "upstream_oauth",
-                        action = "probe",
-                        upstream = %name,
-                        registration_strategy = strategy,
-                        elapsed_ms = started.elapsed().as_millis(),
-                        "upstream oauth probe: transient manager registered"
-                    );
-                }
-            }
-            _ => {
-                tracing::warn!(
-                    service = "upstream_oauth",
-                    action = "probe",
-                    upstream = %name,
-                    kind = "not_configured",
-                    elapsed_ms = started.elapsed().as_millis(),
-                    "upstream oauth probe: oauth resources not configured (LAB_PUBLIC_URL + LAB_OAUTH_ENCRYPTION_KEY required)"
-                );
-                return Err(ToolError::Sdk {
-                    sdk_kind: "not_configured".to_string(),
-                    message: "upstream OAuth requires LAB_PUBLIC_URL (https) and LAB_OAUTH_ENCRYPTION_KEY to be set".to_string(),
-                });
-            }
-        }
-
-        Ok(crate::dispatch::gateway::oauth::ProbeResult {
-            upstream: name,
-            url: url.to_string(),
-            oauth_discovered: true,
-            issuer: metadata.issuer,
-            scopes: metadata.scopes_supported,
-            registration_strategy: Some(strategy.to_string()),
-        })
-    }
-
     /// Returns the upstream OAuth SQLite store, if configured.
-    pub fn oauth_sqlite(&self) -> Option<lab_auth::sqlite::SqliteStore> {
-        self.oauth_sqlite.clone()
-    }
-
     /// Returns the upstream OAuth callback redirect URI, if configured.
     ///
     /// Used by the `/.well-known/oauth-client` endpoint to build the client
     /// metadata document served to CIMD-supporting authorization servers.
-    pub fn oauth_redirect_uri(&self) -> Option<String> {
-        self.oauth_redirect_uri.as_deref().map(|s| s.to_string())
-    }
-
     pub fn upstream_oauth_manager(&self, upstream: &str) -> Option<UpstreamOauthManager> {
         self.upstream_oauth_managers
             .as_ref()
             .and_then(|managers| managers.get(upstream).map(|entry| entry.clone()))
-    }
-
-    pub fn evict_subject_client(&self, upstream: &str, subject: &str) {
-        if let Some(cache) = &self.oauth_client_cache {
-            cache.evict_subject(upstream, subject);
-        }
     }
 
     #[allow(dead_code)]
@@ -563,235 +230,6 @@ impl GatewayManager {
         if let Some(cache) = &self.oauth_client_cache {
             cache.evict_upstream(upstream);
         }
-    }
-
-    pub async fn begin_upstream_authorization(
-        &self,
-        upstream: &str,
-        subject: &str,
-    ) -> Result<BeginAuthorization, ToolError> {
-        let started = std::time::Instant::now();
-        let manager = self.upstream_oauth_manager(upstream).ok_or_else(|| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "start",
-                upstream,
-                kind = "not_found",
-                "upstream oauth start: upstream not found or has no oauth config"
-            );
-            ToolError::Sdk {
-                sdk_kind: "not_found".to_string(),
-                message: format!("upstream '{upstream}' not found or has no oauth config"),
-            }
-        })?;
-
-        let result = manager.begin_authorization(subject).await.map_err(|e| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "start",
-                upstream,
-                kind = e.kind(),
-                elapsed_ms = started.elapsed().as_millis(),
-                "upstream oauth start: begin authorization failed"
-            );
-            tool_error_from_oauth(e)
-        })?;
-
-        tracing::info!(
-            service = "upstream_oauth",
-            action = "start",
-            upstream,
-            elapsed_ms = started.elapsed().as_millis(),
-            "upstream oauth start: authorization URL generated"
-        );
-        Ok(result)
-    }
-
-    pub async fn complete_upstream_authorization_callback(
-        &self,
-        upstream: &str,
-        subject: &str,
-        code: &str,
-        state: &str,
-    ) -> Result<(), ToolError> {
-        let started = std::time::Instant::now();
-        let manager = self.upstream_oauth_manager(upstream).ok_or_else(|| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "callback",
-                upstream,
-                kind = "not_found",
-                "upstream oauth callback: upstream not found or has no oauth config"
-            );
-            ToolError::Sdk {
-                sdk_kind: "not_found".to_string(),
-                message: format!("upstream '{upstream}' not found or has no oauth config"),
-            }
-        })?;
-
-        manager
-            .complete_authorization_callback(subject, code, state)
-            .await
-            .map_err(|e| {
-                tracing::warn!(
-                    service = "upstream_oauth",
-                    action = "callback",
-                    upstream,
-                    kind = e.kind(),
-                    elapsed_ms = started.elapsed().as_millis(),
-                    "upstream oauth callback: token exchange failed"
-                );
-                tool_error_from_oauth(e)
-            })?;
-
-        tracing::info!(
-            service = "upstream_oauth",
-            action = "callback",
-            upstream,
-            elapsed_ms = started.elapsed().as_millis(),
-            "upstream oauth callback: tokens stored"
-        );
-
-        if let Some(oauth_config) = manager.upstream_config().oauth.clone() {
-            let manager_for_persist = self.clone();
-            let upstream_for_persist = upstream.to_string();
-            tokio::spawn(async move {
-                let _mutation_guard = manager_for_persist.config_mutation.lock().await;
-                let mut cfg = manager_for_persist.config.read().await.clone();
-                let Some(existing) = cfg
-                    .upstream
-                    .iter_mut()
-                    .find(|u| u.name == upstream_for_persist.as_str())
-                else {
-                    tracing::debug!(
-                        service = "upstream_oauth",
-                        action = "callback",
-                        upstream = %upstream_for_persist,
-                        "upstream oauth callback: no matching gateway in config; skipping oauth persistence"
-                    );
-                    return;
-                };
-                if existing.oauth.is_some() {
-                    return;
-                }
-                tracing::info!(
-                    service = "upstream_oauth",
-                    action = "callback",
-                    upstream = %upstream_for_persist,
-                    "upstream oauth callback: persisting oauth config for probe-created manager"
-                );
-                existing.oauth = Some(oauth_config);
-                if let Err(e) = manager_for_persist.persist_config(cfg).await {
-                    tracing::warn!(
-                        service = "upstream_oauth",
-                        action = "callback",
-                        upstream = upstream_for_persist,
-                        error = %e,
-                        "upstream oauth callback: failed to persist oauth config (non-fatal)"
-                    );
-                }
-            });
-        }
-
-        Ok(())
-    }
-
-    pub async fn upstream_oauth_status(
-        &self,
-        upstream: &str,
-        subject: &str,
-    ) -> Result<UpstreamOauthStatusView, ToolError> {
-        let started = std::time::Instant::now();
-        let manager = self.upstream_oauth_manager(upstream).ok_or_else(|| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "status",
-                upstream,
-                kind = "not_found",
-                "upstream oauth status: upstream not found or has no oauth config"
-            );
-            ToolError::Sdk {
-                sdk_kind: "not_found".to_string(),
-                message: format!("upstream '{upstream}' not found or has no oauth config"),
-            }
-        })?;
-
-        let row = manager.credential_row(subject).await.map_err(|e| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "status",
-                upstream,
-                kind = e.kind(),
-                elapsed_ms = started.elapsed().as_millis(),
-                "upstream oauth status: credential lookup failed"
-            );
-            tool_error_from_oauth(e)
-        })?;
-
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs() as i64;
-        let authenticated = row.is_some();
-        let expires_within_5m = row.is_some_and(|row| row.access_token_expires_at - now <= 300);
-
-        tracing::debug!(
-            service = "upstream_oauth",
-            action = "status",
-            upstream,
-            authenticated,
-            expires_within_5m,
-            elapsed_ms = started.elapsed().as_millis(),
-            "upstream oauth status: checked"
-        );
-        Ok(UpstreamOauthStatusView {
-            authenticated,
-            upstream: upstream.to_string(),
-            expires_within_5m,
-        })
-    }
-
-    pub async fn clear_upstream_credentials(
-        &self,
-        upstream: &str,
-        subject: &str,
-    ) -> Result<(), ToolError> {
-        let started = std::time::Instant::now();
-        let manager = self.upstream_oauth_manager(upstream).ok_or_else(|| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "clear",
-                upstream,
-                kind = "not_found",
-                "upstream oauth clear: upstream not found or has no oauth config"
-            );
-            ToolError::Sdk {
-                sdk_kind: "not_found".to_string(),
-                message: format!("upstream '{upstream}' not found or has no oauth config"),
-            }
-        })?;
-
-        manager.clear_credentials(subject).await.map_err(|e| {
-            tracing::warn!(
-                service = "upstream_oauth",
-                action = "clear",
-                upstream,
-                kind = e.kind(),
-                elapsed_ms = started.elapsed().as_millis(),
-                "upstream oauth clear: failed to clear credentials"
-            );
-            tool_error_from_oauth(e)
-        })?;
-
-        self.evict_subject_client(upstream, subject);
-        tracing::info!(
-            service = "upstream_oauth",
-            action = "clear",
-            upstream,
-            elapsed_ms = started.elapsed().as_millis(),
-            "upstream oauth clear: credentials cleared and client cache evicted"
-        );
-        Ok(())
     }
 
     pub async fn get_service_config(&self, service: &str) -> Result<ServiceConfigView, ToolError> {
@@ -827,6 +265,7 @@ impl GatewayManager {
             }
         }
 
+        let _mutation_guard = self.config_mutation.lock().await;
         let creds = values_to_service_creds(service, values);
         let env_path = self.env_path();
         if !creds.is_empty() && !env_is_up_to_date(&env_path, &creds) {
@@ -2250,7 +1689,7 @@ impl GatewayManager {
         Ok(())
     }
 
-    async fn persist_config(&self, cfg: LabConfig) -> Result<(), ToolError> {
+    pub(super) async fn persist_config(&self, cfg: LabConfig) -> Result<(), ToolError> {
         let path = self.path.clone();
         let cfg_clone = cfg.clone();
         tracing::info!(
@@ -2270,99 +1709,6 @@ impl GatewayManager {
             "gateway reconcile"
         );
         Ok(())
-    }
-
-    fn runtime_state_path(&self) -> PathBuf {
-        let parent = self
-            .path
-            .parent()
-            .map(ToOwned::to_owned)
-            .unwrap_or_else(|| PathBuf::from("."));
-        let stem = self
-            .path
-            .file_stem()
-            .and_then(|value| value.to_str())
-            .unwrap_or("gateway");
-        parent.join(format!("{stem}.runtime.json"))
-    }
-
-    async fn load_runtime_state(&self) -> PersistedGatewayRuntimeState {
-        let path = self.runtime_state_path();
-        let Ok(raw) = tokio::fs::read_to_string(path).await else {
-            return PersistedGatewayRuntimeState::default();
-        };
-        serde_json::from_str(&raw).unwrap_or_default()
-    }
-
-    async fn persist_runtime_state(
-        &self,
-        state: &PersistedGatewayRuntimeState,
-    ) -> Result<(), ToolError> {
-        let path = self.runtime_state_path();
-        if let Some(parent) = path.parent() {
-            tokio::fs::create_dir_all(parent).await.map_err(|error| {
-                ToolError::internal_message(format!(
-                    "failed to create runtime state directory {}: {error}",
-                    parent.display()
-                ))
-            })?;
-        }
-        let body = serde_json::to_vec_pretty(state).map_err(|error| {
-            ToolError::internal_message(format!("failed to serialize runtime state: {error}"))
-        })?;
-        tokio::fs::write(&path, body).await.map_err(|error| {
-            ToolError::internal_message(format!(
-                "failed to write runtime state {}: {error}",
-                path.display()
-            ))
-        })
-    }
-
-    async fn reconcile_runtime_state(
-        &self,
-        cfg: &LabConfig,
-        pool: Option<&UpstreamPool>,
-    ) -> Result<PersistedGatewayRuntimeState, ToolError> {
-        let mut state = self.load_runtime_state().await;
-        state.entries.retain(|entry| process_is_alive(entry.pid));
-
-        if let Some(pool) = pool {
-            for upstream in &cfg.upstream {
-                if let Some(runtime) = pool.upstream_runtime_metadata(&upstream.name).await
-                    && let Some(pid) = runtime.pid
-                {
-                    state
-                        .entries
-                        .retain(|entry| !(entry.upstream == upstream.name && entry.pid == pid));
-                    state.entries.push(PersistedGatewayRuntimeEntry {
-                        upstream: upstream.name.clone(),
-                        pid,
-                        pgid: runtime.pgid,
-                        started_at_epoch_secs: runtime
-                            .started_at
-                            .and_then(system_time_to_epoch_secs),
-                        observed_at_epoch_secs: epoch_now_secs(),
-                        origin: runtime.origin.clone(),
-                        owner: runtime.owner.as_ref().map(runtime_owner_view),
-                        transport: Some(if upstream.command.is_some() {
-                            "stdio".to_string()
-                        } else {
-                            "http".to_string()
-                        }),
-                        target: redacted_gateway_target(upstream),
-                    });
-                }
-            }
-        }
-
-        state.reconciled_at_epoch_secs = Some(epoch_now_secs());
-        state.entries.sort_by(|left, right| {
-            left.upstream
-                .cmp(&right.upstream)
-                .then(left.pid.cmp(&right.pid))
-        });
-        self.persist_runtime_state(&state).await?;
-        Ok(state)
     }
 }
 
@@ -2439,767 +1785,8 @@ fn quarantine_unregistered_virtual_servers(
     (cfg, migration)
 }
 
-fn config_view(upstream: &UpstreamConfig, tool_search: &ToolSearchConfig) -> GatewayConfigView {
-    GatewayConfigView {
-        name: upstream.name.clone(),
-        enabled: upstream.enabled,
-        url: upstream.url.as_deref().map(redact_url),
-        command: upstream.command.as_deref().map(redact_stdio_value),
-        args: redact_stdio_args(&upstream.args),
-        bearer_token_env: upstream.bearer_token_env.clone(),
-        oauth_enabled: upstream.oauth.is_some(),
-        proxy_resources: upstream.proxy_resources,
-        proxy_prompts: upstream.proxy_prompts,
-        expose_tools: upstream.expose_tools.clone(),
-        expose_resources: upstream.expose_resources.clone(),
-        expose_prompts: upstream.expose_prompts.clone(),
-        tool_search_enabled: tool_search.enabled,
-        tool_search_top_k_default: tool_search.top_k_default,
-        tool_search_max_tools: tool_search.max_tools,
-    }
-}
-
-fn sanitize_tool_text(input: &str, max_len: usize) -> String {
-    let mut sanitized = input.to_string();
-    sanitized.retain(|ch| {
-        !matches!(
-            ch,
-            '\u{0000}'..='\u{001F}'
-                | '\u{007F}'..='\u{009F}'
-                | '\u{202A}'..='\u{202E}'
-                | '\u{2066}'..='\u{2069}'
-        )
-    });
-    for marker in ["<system>", "[INST]", "###", "<<"] {
-        sanitized = sanitized.replace(marker, "");
-    }
-    redact_secret_like_segments(&sanitized)
-        .chars()
-        .take(max_len)
-        .collect()
-}
-
-fn redact_secret_like_segments(input: &str) -> String {
-    input
-        .split_whitespace()
-        .map(|segment| {
-            let looks_secret = segment.starts_with("sk-")
-                || segment.starts_with("ghp_")
-                || segment.starts_with("github_pat_")
-                || segment.starts_with("glpat-")
-                || segment.starts_with("xoxb-")
-                || segment.starts_with("xoxp-")
-                || segment.starts_with("eyJ");
-            if looks_secret {
-                "<redacted>".to_string()
-            } else {
-                segment.to_string()
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn sanitize_schema(schema: Option<serde_json::Value>) -> Option<serde_json::Value> {
-    fn recurse(value: &mut serde_json::Value) {
-        match value {
-            serde_json::Value::String(text) => {
-                *text = sanitize_tool_text(text, 2048);
-            }
-            serde_json::Value::Array(values) => {
-                for value in values {
-                    recurse(value);
-                }
-            }
-            serde_json::Value::Object(map) => {
-                for value in map.values_mut() {
-                    recurse(value);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    schema.map(|mut value| {
-        recurse(&mut value);
-        value
-    })
-}
-
-fn redacted_gateway_target(upstream: &UpstreamConfig) -> Option<String> {
-    upstream.url.as_deref().map(redact_url).or_else(|| {
-        upstream.command.as_deref().map(|command| {
-            let args = redact_stdio_args(&upstream.args);
-            format_redacted_gateway_command(command, &args)
-        })
-    })
-}
-
-fn ensure_stdio_admin_ack(
-    action: &str,
-    upstream: &UpstreamConfig,
-    allow_stdio: bool,
-) -> Result<(), ToolError> {
-    if upstream.command.is_none() || !upstream.enabled || allow_stdio {
-        return Ok(());
-    }
-
-    Err(ToolError::InvalidParam {
-        message: format!(
-            "{action} would execute local command `{}` for stdio gateway `{}`; resend with `allow_stdio: true` only after operator approval",
-            upstream.command.as_deref().unwrap_or("<unknown>"),
-            upstream.name
-        ),
-        param: "allow_stdio".to_string(),
-    })
-}
-
-fn format_redacted_gateway_command(command: &str, args: &[String]) -> String {
-    if command == "env" {
-        let _ = args;
-        return "env".to_string();
-    }
-
-    redact_stdio_value(command)
-}
-
-fn empty_upstream_summary() -> UpstreamCachedSummary {
-    UpstreamCachedSummary::default()
-}
-
-fn is_nonessential_capability_error(message: &str) -> bool {
-    // Only suppress the well-known optional-capability discovery failures
-    // (prompts/resources list not implemented). Broad "-32601" / "Method not
-    // found" matching would also hide real tool-call or handshake failures.
-    message.starts_with("failed to list prompts from upstream:")
-        || message.starts_with("failed to list resources from upstream:")
-        || message.starts_with("does not implement MCP prompts discovery")
-        || message.starts_with("does not implement MCP resources discovery")
-}
-
-fn operator_visible_upstream_error(message: Option<String>) -> Option<String> {
-    message.filter(|message| !is_nonessential_capability_error(message))
-}
-
-async fn upstream_summary(
-    pool: Option<&UpstreamPool>,
-    upstream_name: &str,
-) -> UpstreamCachedSummary {
-    let Some(pool) = pool else {
-        return empty_upstream_summary();
-    };
-
-    pool.cached_upstream_summary(upstream_name)
-        .await
-        .unwrap_or_else(empty_upstream_summary)
-}
-
-async fn server_view_from_upstream(
-    pool: Option<&UpstreamPool>,
-    upstream: &UpstreamConfig,
-) -> ServerView {
-    let summary = upstream_summary(pool, &upstream.name).await;
-    let last_error = operator_visible_upstream_error(match pool {
-        Some(pool) => pool.upstream_last_error(&upstream.name).await,
-        None => None,
-    });
-    let connected = summary.exposed_tool_count > 0
-        || summary.exposed_resource_count > 0
-        || summary.exposed_prompt_count > 0;
-    let enabled = upstream.enabled;
-
-    ServerView {
-        id: upstream.name.clone(),
-        name: upstream.name.clone(),
-        source: "custom_gateway".to_string(),
-        configured: true,
-        enabled,
-        connected: enabled && connected,
-        discovered_tool_count: summary.discovered_tool_count,
-        exposed_tool_count: summary.exposed_tool_count,
-        discovered_resource_count: summary.discovered_resource_count,
-        exposed_resource_count: summary.exposed_resource_count,
-        discovered_prompt_count: summary.discovered_prompt_count,
-        exposed_prompt_count: summary.exposed_prompt_count,
-        surfaces: SurfaceStatesView {
-            mcp: SurfaceStateView {
-                enabled,
-                connected: enabled && connected,
-            },
-            ..SurfaceStatesView::default()
-        },
-        warnings: last_error
-            .as_ref()
-            .map(|message| {
-                vec![super::view_models::ServerWarningView {
-                    code: "connection_error".to_string(),
-                    message: message.clone(),
-                }]
-            })
-            .unwrap_or_default(),
-        config_summary: ServerConfigSummaryView {
-            transport: Some(if upstream.command.is_some() {
-                "stdio".to_string()
-            } else {
-                "http".to_string()
-            }),
-            target: redacted_gateway_target(upstream),
-        },
-    }
-}
-
-impl GatewayManager {
-    pub async fn mcp_runtime_list(
-        &self,
-    ) -> Result<Vec<super::types::GatewayMcpRuntimeView>, ToolError> {
-        let cfg = self.config.read().await.clone();
-        let pool = self.runtime.current_pool().await;
-        let persisted = self.reconcile_runtime_state(&cfg, pool.as_deref()).await?;
-        let mut rows = Vec::with_capacity(cfg.upstream.len());
-        for upstream in &cfg.upstream {
-            let summary = upstream_summary(pool.as_deref(), &upstream.name).await;
-            let runtime = match pool.as_deref() {
-                Some(pool) => pool.upstream_runtime_metadata(&upstream.name).await,
-                None => None,
-            };
-            let live_pid = runtime.as_ref().and_then(|meta| meta.pid);
-            let persisted_rows: Vec<&PersistedGatewayRuntimeEntry> = persisted
-                .entries
-                .iter()
-                .filter(|entry| entry.upstream == upstream.name)
-                .collect();
-            let stale_count = persisted_rows
-                .iter()
-                .filter(|entry| Some(entry.pid) != live_pid)
-                .count();
-            let fallback = if let Some(pid) = live_pid {
-                persisted_rows.into_iter().find(|entry| entry.pid == pid)
-            } else {
-                persisted_rows.into_iter().max_by_key(|entry| {
-                    entry
-                        .started_at_epoch_secs
-                        .unwrap_or(entry.observed_at_epoch_secs)
-                })
-            };
-            let connected = upstream.enabled
-                && (summary.exposed_tool_count > 0
-                    || summary.exposed_resource_count > 0
-                    || summary.exposed_prompt_count > 0);
-            rows.push(super::types::GatewayMcpRuntimeView {
-                name: upstream.name.clone(),
-                enabled: upstream.enabled,
-                connected,
-                discovered_tool_count: summary.discovered_tool_count,
-                exposed_tool_count: summary.exposed_tool_count,
-                discovered_resource_count: summary.discovered_resource_count,
-                exposed_resource_count: summary.exposed_resource_count,
-                discovered_prompt_count: summary.discovered_prompt_count,
-                exposed_prompt_count: summary.exposed_prompt_count,
-                likely_stale_count: stale_count,
-                pid: live_pid.or_else(|| fallback.map(|entry| entry.pid)),
-                pgid: runtime
-                    .as_ref()
-                    .and_then(|meta| meta.pgid)
-                    .or_else(|| fallback.and_then(|entry| entry.pgid)),
-                age_seconds: runtime
-                    .as_ref()
-                    .and_then(|meta| meta.started_at)
-                    .and_then(|started_at| {
-                        std::time::SystemTime::now().duration_since(started_at).ok()
-                    })
-                    .map(|elapsed: Duration| elapsed.as_secs())
-                    .or_else(|| {
-                        fallback
-                            .and_then(|entry| entry.started_at_epoch_secs)
-                            .and_then(age_from_epoch_secs)
-                    }),
-                origin: runtime
-                    .as_ref()
-                    .and_then(|meta| meta.origin.clone())
-                    .or_else(|| fallback.and_then(|entry| entry.origin.clone())),
-                owner: runtime
-                    .as_ref()
-                    .and_then(|meta| meta.owner.as_ref().map(runtime_owner_view))
-                    .or_else(|| fallback.and_then(|entry| entry.owner.clone())),
-                transport: Some(if upstream.command.is_some() {
-                    "stdio".to_string()
-                } else {
-                    "http".to_string()
-                }),
-                target: fallback
-                    .and_then(|entry| entry.target.clone())
-                    .or_else(|| redacted_gateway_target(upstream)),
-                runtime_state_path: Some(self.runtime_state_path().display().to_string()),
-                reconciled_at: persisted
-                    .reconciled_at_epoch_secs
-                    .and_then(epoch_secs_to_rfc3339),
-            });
-        }
-        Ok(rows)
-    }
-
-    pub async fn cleanup_upstream_processes(
-        &self,
-        name: &str,
-        aggressive: bool,
-        dry_run: bool,
-    ) -> Result<super::types::GatewayCleanupView, ToolError> {
-        let upstream = self
-            .config
-            .read()
-            .await
-            .upstream
-            .iter()
-            .find(|existing| existing.name == name)
-            .cloned()
-            .ok_or_else(|| ToolError::Sdk {
-                sdk_kind: "not_found".to_string(),
-                message: format!("gateway `{name}` not found"),
-            })?;
-
-        let gateway_patterns = upstream_cleanup_patterns(&upstream, false);
-        let local_patterns = local_cleanup_patterns();
-        let aggressive_patterns = if aggressive {
-            upstream_cleanup_patterns(&upstream, true)
-        } else {
-            Vec::new()
-        };
-
-        let gateway_matches = matching_processes(&gateway_patterns);
-        let local_matches = matching_processes(&local_patterns);
-        let aggressive_matches = if aggressive {
-            matching_processes(&aggressive_patterns)
-        } else {
-            Vec::new()
-        };
-
-        let view = super::types::GatewayCleanupView {
-            upstream: upstream.name,
-            aggressive,
-            dry_run,
-            gateway_matched: count_matched_processes(&gateway_matches),
-            local_matched: count_matched_processes(&local_matches),
-            aggressive_matched: if aggressive {
-                count_matched_processes(&aggressive_matches)
-            } else {
-                0
-            },
-            gateway_killed: if dry_run {
-                count_matched_processes(&gateway_matches)
-            } else {
-                kill_matched_processes(&gateway_matches)
-            },
-            local_killed: if dry_run {
-                count_matched_processes(&local_matches)
-            } else {
-                kill_matched_processes(&local_matches)
-            },
-            aggressive_killed: if aggressive {
-                if dry_run {
-                    count_matched_processes(&aggressive_matches)
-                } else {
-                    kill_matched_processes(&aggressive_matches)
-                }
-            } else {
-                0
-            },
-            gateway_matches: gateway_matches.iter().map(cleanup_match_view).collect(),
-            local_matches: local_matches.iter().map(cleanup_match_view).collect(),
-            aggressive_matches: aggressive_matches.iter().map(cleanup_match_view).collect(),
-        };
-
-        let cfg = self.config.read().await.clone();
-        let current_pool = self.runtime.current_pool().await;
-        self.reconcile_runtime_state(&cfg, current_pool.as_deref())
-            .await?;
-
-        Ok(view)
-    }
-}
-
-fn local_cleanup_patterns() -> Vec<String> {
-    vec![
-        "labby mcp".to_string(),
-        "target/debug/labby mcp".to_string(),
-    ]
-}
-
-fn upstream_cleanup_patterns(upstream: &UpstreamConfig, aggressive: bool) -> Vec<String> {
-    let mut patterns = Vec::new();
-    let command = upstream.command.as_deref().unwrap_or("");
-    let joined_args = upstream.args.join(" ");
-    let joined = if command.is_empty() {
-        joined_args.clone()
-    } else if joined_args.is_empty() {
-        command.to_string()
-    } else {
-        format!("{command} {joined_args}")
-    };
-    if let Some(command) = upstream.command.as_deref() {
-        let mut joined = command.to_string();
-        for arg in &upstream.args {
-            joined.push(' ');
-            joined.push_str(arg);
-        }
-        patterns.push(joined);
-        for arg in &upstream.args {
-            if arg.contains("mcp") || arg.contains(&upstream.name) {
-                patterns.push(arg.clone());
-            }
-        }
-    }
-    if joined.contains("chrome-devtools-mcp") || upstream.name.contains("chrome-devtools") {
-        patterns.push("chrome-devtools-mcp".to_string());
-        patterns.push("chrome-devtools".to_string());
-        patterns.push("chrome-devtools-mcp/build/src/telemetry/watchdog/main.js".to_string());
-        patterns.push("npm exec chrome-devtools-mcp@latest".to_string());
-    }
-    if joined.contains("github-chat-mcp") || upstream.name.contains("github-chat") {
-        patterns.push("github-chat-mcp".to_string());
-        patterns.push("uvx github-chat-mcp".to_string());
-        patterns.push("uv tool uvx github-chat-mcp".to_string());
-        patterns.push("uv run github-chat-mcp".to_string());
-        patterns.push("github-chat".to_string());
-        patterns.push("/github-chat-mcp".to_string());
-    }
-    if aggressive {
-        patterns.push(upstream.name.clone());
-    }
-    patterns.sort();
-    patterns.dedup();
-    patterns
-}
-
-fn epoch_now_secs() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|elapsed| elapsed.as_secs())
-        .unwrap_or(0)
-}
-
-fn system_time_to_epoch_secs(time: std::time::SystemTime) -> Option<u64> {
-    time.duration_since(std::time::UNIX_EPOCH)
-        .ok()
-        .map(|elapsed| elapsed.as_secs())
-}
-
-fn age_from_epoch_secs(epoch_secs: u64) -> Option<u64> {
-    let started_at = std::time::UNIX_EPOCH.checked_add(Duration::from_secs(epoch_secs))?;
-    std::time::SystemTime::now()
-        .duration_since(started_at)
-        .ok()
-        .map(|elapsed| elapsed.as_secs())
-}
-
-fn epoch_secs_to_rfc3339(epoch_secs: u64) -> Option<String> {
-    let seconds = i64::try_from(epoch_secs).ok()?;
-    let timestamp = jiff::Timestamp::from_second(seconds).ok()?;
-    Some(timestamp.to_string())
-}
-
-fn runtime_origin_tag(origin: Option<&str>) -> Option<String> {
-    origin
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn runtime_owner_view(
-    owner: &UpstreamRuntimeOwner,
-) -> crate::dispatch::gateway::types::GatewayRuntimeOwnerView {
-    crate::dispatch::gateway::types::GatewayRuntimeOwnerView {
-        surface: owner.surface.clone(),
-        subject: owner.subject.clone(),
-        request_id: owner.request_id.clone(),
-        session_id: owner.session_id.clone(),
-        client_name: owner.client_name.clone(),
-        raw: owner.raw.clone(),
-    }
-}
-
-#[cfg(unix)]
-fn process_is_alive(pid: u32) -> bool {
-    pid_is_alive(pid)
-}
-
-#[cfg(not(unix))]
-fn process_is_alive(_pid: u32) -> bool {
-    false
-}
-
-#[derive(Debug, Clone, Default)]
-struct GatewayCleanupMatch {
-    pattern: String,
-    pids: Vec<u32>,
-}
-
-fn cleanup_match_view(matched: &GatewayCleanupMatch) -> super::types::GatewayCleanupMatchView {
-    super::types::GatewayCleanupMatchView {
-        pattern: matched.pattern.clone(),
-        pids: matched.pids.clone(),
-    }
-}
-
-#[cfg(target_os = "linux")]
-fn current_and_parent_pids() -> std::collections::HashSet<u32> {
-    let mut pids = std::collections::HashSet::from([std::process::id()]);
-    let parent = nix::unistd::getppid();
-    if parent.as_raw() > 0 {
-        pids.insert(parent.as_raw() as u32);
-    }
-    pids
-}
-
-#[cfg(target_os = "linux")]
-fn matching_processes(patterns: &[String]) -> Vec<GatewayCleanupMatch> {
-    let excluded_pids = current_and_parent_pids();
-    let mut matched: BTreeMap<String, BTreeSet<u32>> = BTreeMap::new();
-    for entry in std::fs::read_dir("/proc")
-        .ok()
-        .into_iter()
-        .flatten()
-        .flatten()
-    {
-        let pid_str = entry.file_name();
-        let Ok(pid) = pid_str.to_string_lossy().parse::<u32>() else {
-            continue;
-        };
-        if excluded_pids.contains(&pid) {
-            continue;
-        }
-        let Some(cmdline) = read_cmdline(pid) else {
-            continue;
-        };
-        for pattern in patterns.iter().filter(|pattern| !pattern.trim().is_empty()) {
-            if cmdline.contains(pattern) {
-                matched.entry(pattern.clone()).or_default().insert(pid);
-            }
-        }
-    }
-    matched
-        .into_iter()
-        .map(|(pattern, pids)| GatewayCleanupMatch {
-            pattern,
-            pids: pids.into_iter().collect(),
-        })
-        .collect()
-}
-
-#[cfg(not(target_os = "linux"))]
-fn matching_processes(_patterns: &[String]) -> Vec<GatewayCleanupMatch> {
-    Vec::new()
-}
-
-#[cfg(all(test, target_os = "linux"))]
-fn process_matches_patterns(cmdline: &str, patterns: &[String]) -> bool {
-    patterns
-        .iter()
-        .filter(|pattern| !pattern.trim().is_empty())
-        .any(|pattern| cmdline.contains(pattern))
-}
-
-fn count_matched_processes(matches: &[GatewayCleanupMatch]) -> usize {
-    let mut unique = BTreeSet::new();
-    for matched in matches {
-        unique.extend(matched.pids.iter().copied());
-    }
-    unique.len()
-}
-
-fn kill_matched_processes(matches: &[GatewayCleanupMatch]) -> usize {
-    let mut unique = BTreeSet::new();
-    for matched in matches {
-        unique.extend(matched.pids.iter().copied());
-    }
-    for pid in &unique {
-        let _ = terminate_process(*pid);
-    }
-    unique.len()
-}
-
-#[cfg(unix)]
-fn terminate_process(pid: u32) -> Result<(), Errno> {
-    terminate_sigkill(pid)
-}
-
-#[cfg(not(unix))]
-fn terminate_process(_pid: u32) -> Result<(), ()> {
-    Ok(())
-}
-
-fn degraded_server_warning_count(views: &[ServerView], code: &str) -> usize {
-    views
-        .iter()
-        .filter(|view| view.warnings.iter().any(|warning| warning.code == code))
-        .count()
-}
-
-fn server_view_from_virtual_server(
-    config: &crate::config::VirtualServerConfig,
-    summary: UpstreamCachedSummary,
-    last_error: Option<String>,
-    health: Option<&ServiceHealth>,
-) -> ServerView {
-    let record = VirtualServerRecord::from(config);
-    let service = match &record.source {
-        VirtualServerSource::LabService { service } => service.clone(),
-    };
-    let service_known = service_meta(&service).is_some();
-    let peer_connected = last_error.is_none()
-        && (summary.discovered_tool_count > 0
-            || summary.discovered_resource_count > 0
-            || summary.discovered_prompt_count > 0);
-    let health_connected = health
-        .map(|health| health.reachable && health.auth_ok)
-        .unwrap_or(false);
-    let connected = service_known && record.enabled && (peer_connected || health_connected);
-    let mcp_exposed = record.enabled && record.surfaces.mcp;
-    let discovered_tool_count = summary.discovered_tool_count;
-    let policy_exposed_tool_count = record.mcp_policy.as_ref().and_then(|policy| {
-        (!policy.allowed_actions.is_empty()).then_some(policy.allowed_actions.len() + 2)
-    });
-    let exposed_tool_count = if mcp_exposed {
-        policy_exposed_tool_count
-            .map(|count| summary.exposed_tool_count.min(count))
-            .unwrap_or(summary.exposed_tool_count)
-    } else {
-        0
-    };
-    let discovered_resource_count = summary.discovered_resource_count;
-    let exposed_resource_count = if mcp_exposed {
-        summary.exposed_resource_count
-    } else {
-        0
-    };
-    let discovered_prompt_count = summary.discovered_prompt_count;
-    let exposed_prompt_count = if mcp_exposed {
-        summary.exposed_prompt_count
-    } else {
-        0
-    };
-    let mut warnings = Vec::new();
-    if !service_known {
-        warnings.push(super::view_models::ServerWarningView {
-            code: WARNING_UNKNOWN_SERVICE.to_string(),
-            message: format!("service `{service}` is not registered in this lab binary"),
-        });
-    }
-    if let Some(message) = last_error {
-        warnings.push(super::view_models::ServerWarningView {
-            code: "connection_error".to_string(),
-            message,
-        });
-    }
-
-    ServerView {
-        id: record.id.clone(),
-        name: service.clone(),
-        source: "in_process".to_string(),
-        configured: true,
-        enabled: record.enabled,
-        connected,
-        discovered_tool_count,
-        exposed_tool_count,
-        discovered_resource_count,
-        exposed_resource_count,
-        discovered_prompt_count,
-        exposed_prompt_count,
-        surfaces: SurfaceStatesView {
-            cli: SurfaceStateView {
-                enabled: record.surfaces.cli,
-                connected: record.surfaces.cli && connected,
-            },
-            api: SurfaceStateView {
-                enabled: record.surfaces.api,
-                connected: record.surfaces.api && connected,
-            },
-            mcp: SurfaceStateView {
-                enabled: record.surfaces.mcp,
-                connected: record.surfaces.mcp && connected,
-            },
-            webui: SurfaceStateView {
-                enabled: record.surfaces.webui,
-                connected: record.surfaces.webui && connected,
-            },
-        },
-        warnings,
-        config_summary: ServerConfigSummaryView {
-            transport: Some("in_process".to_string()),
-            target: Some(service),
-        },
-    }
-}
-
 fn builtin_service_registry() -> &'static ToolRegistry {
     BUILTIN_SERVICE_REGISTRY.get_or_init(crate::registry::build_default_registry)
-}
-
-fn read_env_values(path: &std::path::Path) -> Result<HashMap<String, String>, ToolError> {
-    Ok(dotenvy::from_path_iter(path)
-        .ok()
-        .map(|iter| iter.filter_map(Result::ok).collect())
-        .unwrap_or_default())
-}
-
-fn values_to_service_creds(service: &str, values: &BTreeMap<String, String>) -> Vec<ServiceCreds> {
-    values
-        .iter()
-        .map(|(field, value)| {
-            let url = if field == &format!("{}_URL", service.to_uppercase()) {
-                Some(value.clone())
-            } else {
-                None
-            };
-            let secret = if url.is_some() {
-                None
-            } else {
-                Some(value.clone())
-            };
-            ServiceCreds {
-                service: service.to_string(),
-                url,
-                secret,
-                env_field: field.clone(),
-                source_host: None,
-                probe_host: None,
-                runtime: None,
-                url_verified: false,
-            }
-        })
-        .collect()
-}
-
-fn service_config_view(
-    meta: &lab_apis::core::PluginMeta,
-    values: &HashMap<String, String>,
-) -> ServiceConfigView {
-    let mut fields = Vec::new();
-    for env in meta.required_env.iter().chain(meta.optional_env.iter()) {
-        let value = values
-            .get(env.name)
-            .and_then(|value| (!value.trim().is_empty()).then_some(value));
-        fields.push(ServiceConfigFieldView {
-            name: env.name.to_string(),
-            present: value.is_some(),
-            secret: env.secret,
-            value_preview: value.and_then(|value| (!env.secret).then(|| value.clone())),
-        });
-    }
-
-    ServiceConfigView {
-        service: meta.name.to_string(),
-        // A service with no env vars needs no configuration and is always ready.
-        configured: if fields.is_empty() {
-            true
-        } else {
-            meta.required_env.iter().all(|env| {
-                fields
-                    .iter()
-                    .any(|field| field.name == env.name && field.present)
-            })
-        },
-        fields,
-    }
 }
 
 async fn snapshot_from_pool(pool: Option<Arc<UpstreamPool>>) -> GatewayCatalogSnapshot {
@@ -3229,46 +1816,20 @@ async fn snapshot_from_pool(pool: Option<Arc<UpstreamPool>>) -> GatewayCatalogSn
     }
 }
 
-async fn runtime_view(
-    pool: Option<&UpstreamPool>,
-    name: &str,
-    prompt_owners: Option<&HashMap<String, String>>,
-) -> GatewayRuntimeView {
-    let Some(pool) = pool else {
-        return GatewayRuntimeView {
-            name: name.to_string(),
-            ..GatewayRuntimeView::default()
-        };
-    };
-
-    let summary = pool
-        .cached_upstream_summary(name)
-        .await
-        .unwrap_or_else(empty_upstream_summary);
-    let prompt_count = match prompt_owners {
-        Some(owners) => owners.values().filter(|owner| *owner == name).count(),
-        None => summary.exposed_prompt_count,
-    };
-
-    GatewayRuntimeView {
-        name: name.to_string(),
-        tool_count: summary.discovered_tool_count,
-        resource_count: summary.discovered_resource_count,
-        prompt_count,
-        exposed_tool_count: summary.exposed_tool_count,
-        exposed_resource_count: summary.exposed_resource_count,
-        exposed_prompt_count: summary.exposed_prompt_count,
-        last_error: operator_visible_upstream_error(pool.upstream_last_error(name).await),
-    }
-}
+#[cfg(test)]
+use super::runtime::{process_matches_patterns, upstream_cleanup_patterns};
 
 #[cfg(test)]
 mod tests {
-    use std::collections::BTreeSet;
+    use std::collections::{BTreeSet, HashMap};
     use std::sync::Arc;
 
-    use crate::config::{UpstreamConfig, VirtualServerConfig, VirtualServerSurfacesConfig};
+    use crate::config::{
+        UpstreamConfig, UpstreamOauthConfig, UpstreamOauthMode, UpstreamOauthRegistration,
+        VirtualServerConfig, VirtualServerSurfacesConfig,
+    };
     use crate::oauth::upstream::cache::OauthClientCache;
+    use crate::tui::events::ServiceHealth;
     use rmcp::transport::{AuthClient, AuthorizationManager};
 
     use super::*;
