+++ crates/lab/src/mcp/catalog.rs
@@ -79,10 +79,12 @@ impl LabMcpServer {
         let mut catalog = crate::catalog::build_catalog(&self.registry);
         let mut services = Vec::new();
         for mut service in catalog.services {
-            if !self.service_visible_on_mcp(&service.name).await {
+            let visible_on_mcp = self.service_visible_on_mcp(&service.name).await;
+            if !visible_on_mcp {
                 continue;
             }
-            if !crate::registry::lab_show_all_enabled()
+            if self.gateway_manager.is_none()
+                && !crate::registry::lab_show_all_enabled()
                 && !crate::registry::service_visible_with_env(&service.name)
             {
                 continue;
@@ -104,7 +106,8 @@ impl LabMcpServer {
         if !self.service_visible_on_mcp(service).await {
             anyhow::bail!("unknown service: {service}");
         }
-        if !crate::registry::lab_show_all_enabled()
+        if self.gateway_manager.is_none()
+            && !crate::registry::lab_show_all_enabled()
             && !crate::registry::service_visible_with_env(service)
         {
             anyhow::bail!("unknown service: {service}");
